% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{rembg-package}
\alias{rembg}
\alias{rembg-package}
\title{\packageTitle{rembg}}
\description{\packageDescription{rembg}}
\details{
On first use the ONNX Runtime shared library must be present. If \pkg{onnxr}
cannot find it, run \code{onnxr::onnx_install()} once.

}
\section{Package Content}{\packageIndices{rembg}}
\author{\packageAuthor{rembg}}
\section{Maintainer}{\packageMaintainer{rembg}}
\keyword{internal}
