% tinyrox says don't edit this manually, but it can't stop you!
\name{rembg_models}
\alias{rembg_models}
\title{Available background-removal models}
\usage{
rembg_models()
}
\value{
A character vector of model names.
}
\description{
Returns the names of the segmentation models that [new_session()] and
[rembg()] can use. Models are downloaded on first use.
}
\examples{
rembg_models()
}
